requirements.txt
Pillow==5.2.0
pandas==0.24.0
numpy==1.14.5
matplotlib==2.2.2
easydict==1.9
scikit-image==0.14.0
scikit-learn==0.20.2
seaborn==0.9.0
jupyter==1.0.0
tensorflow-gpu==1.4.0rc1
tensorflow-tensorboard==0.4.0
Keras==2.1.3